Remove ability to use internal callbacks
