+++ src/mx_bluesky/hyperion/__main__.py
@@ -37,12 +37,6 @@ from mx_bluesky.hyperion.experiment_plans.experiment_registry import (
     PLAN_REGISTRY,
     PlanNotFound,
 )
-from mx_bluesky.hyperion.external_interaction.callbacks.__main__ import (
-    setup_logging as setup_callback_logging,
-)
-from mx_bluesky.hyperion.external_interaction.callbacks.common.callback_util import (
-    CallbacksFactory,
-)
 from mx_bluesky.hyperion.parameters.cli import parse_cli_args
 from mx_bluesky.hyperion.parameters.constants import CONST
 from mx_bluesky.hyperion.utils.context import setup_context
@@ -56,7 +50,6 @@ class Command:
     devices: Any | None = None
     experiment: Callable[[Any, Any], MsgGenerator] | None = None
     parameters: MxBlueskyParameters | None = None
-    callbacks: CallbacksFactory | None = None
 
 
 @dataclass
@@ -88,7 +81,6 @@ class BlueskyRunner:
         RE: RunEngine,
         context: BlueskyContext,
         skip_startup_connection=False,
-        use_external_callbacks: bool = False,
     ) -> None:
         self.command_queue: Queue[Command] = Queue()
         self.current_status: StatusAndMessage = StatusAndMessage(Status.IDLE)
@@ -103,11 +95,9 @@ class BlueskyRunner:
         RE.subscribe(self.aperture_change_callback)
         RE.subscribe(self.logging_uid_tag_callback)
 
-        self.use_external_callbacks = use_external_callbacks
-        if self.use_external_callbacks:
-            LOGGER.info("Connecting to external callback ZMQ proxy...")
-            self.publisher = Publisher(f"localhost:{CONST.CALLBACK_0MQ_PROXY_PORTS[0]}")
-            RE.subscribe(self.publisher)
+        LOGGER.info("Connecting to external callback ZMQ proxy...")
+        self.publisher = Publisher(f"localhost:{CONST.CALLBACK_0MQ_PROXY_PORTS[0]}")
+        RE.subscribe(self.publisher)
 
         if VERBOSE_EVENT_LOGGING:
             RE.subscribe(VerbosePlanExecutionLoggingCallback())
@@ -123,7 +113,6 @@ class BlueskyRunner:
         experiment: Callable,
         parameters: MxBlueskyParameters,
         plan_name: str,
-        callbacks: CallbacksFactory | None,
     ) -> StatusAndMessage:
         LOGGER.info(f"Started with parameters: {parameters.model_dump_json(indent=2)}")
 
@@ -142,7 +131,6 @@ class BlueskyRunner:
                     devices=devices,
                     experiment=experiment,
                     parameters=parameters,
-                    callbacks=callbacks,
                 )
             )
             return StatusAndMessage(Status.SUCCESS)
@@ -181,17 +169,6 @@ class BlueskyRunner:
                 if command.experiment is None:
                     raise ValueError("No experiment provided for START")
                 try:
-                    if (
-                        not self.use_external_callbacks
-                        and command.callbacks
-                        and (cbs := command.callbacks())
-                    ):
-                        LOGGER.info(
-                            f"Using callbacks for this plan: {not self.use_external_callbacks} - {cbs}"
-                        )
-                        self.subscribed_per_plan_callbacks += [
-                            self.RE.subscribe(cb) for cb in cbs
-                        ]
                     with TRACER.start_span("do_run"):
                         self.RE(command.experiment(command.devices, command.parameters))
 
@@ -225,7 +202,6 @@ def compose_start_args(context: BlueskyContext, plan_name: str, action: Actions)
         raise PlanNotFound(f"Experiment plan '{plan_name}' not found in registry.")
 
     experiment_internal_param_type = experiment_registry_entry.get("param_type")
-    callback_type = experiment_registry_entry.get("callback_collection_type")
     plan = context.plan_functions.get(plan_name)
     if experiment_internal_param_type is None:
         raise PlanNotFound(
@@ -243,7 +219,7 @@ def compose_start_args(context: BlueskyContext, plan_name: str, action: Actions)
         raise ValueError(
             f"Supplied parameters don't match the plan for this endpoint {request.data}, for plan {plan_name}"
         ) from e
-    return plan, parameters, plan_name, callback_type
+    return plan, parameters, plan_name
 
 
 class RunExperiment(Resource):
@@ -256,12 +232,10 @@ class RunExperiment(Resource):
         status_and_message = StatusAndMessage(Status.FAILED, f"{action} not understood")
         if action == Actions.START.value:
             try:
-                plan, params, plan_name, callback_type = compose_start_args(
+                plan, params, plan_name = compose_start_args(
                     self.context, plan_name, action
                 )
-                status_and_message = self.runner.start(
-                    plan, params, plan_name, callback_type
-                )
+                status_and_message = self.runner.start(plan, params, plan_name)
             except Exception as e:
                 status_and_message = make_error_status_and_message(e)
                 LOGGER.error(format_exception(e))
@@ -312,7 +286,6 @@ def create_app(
     test_config=None,
     RE: RunEngine = RunEngine({}),
     skip_startup_connection: bool = False,
-    use_external_callbacks: bool = False,
 ) -> tuple[Flask, BlueskyRunner]:
     context = setup_context(
         wait_for_connection=not skip_startup_connection,
@@ -320,7 +293,6 @@ def create_app(
     runner = BlueskyRunner(
         RE,
         context=context,
-        use_external_callbacks=use_external_callbacks,
         skip_startup_connection=skip_startup_connection,
     )
     app = Flask(__name__)
@@ -350,11 +322,8 @@ def create_targets():
     do_default_logging_setup(
         CONST.LOG_FILE_NAME, CONST.GRAYLOG_PORT, dev_mode=args.dev_mode
     )
-    if not args.use_external_callbacks:
-        setup_callback_logging(args.dev_mode)
     app, runner = create_app(
         skip_startup_connection=args.skip_startup_connection,
-        use_external_callbacks=args.use_external_callbacks,
     )
     return app, runner, hyperion_port, args.dev_mode
