+++ src/mx_bluesky/hyperion/experiment_plans/experiment_registry.py
@@ -11,13 +11,6 @@ from mx_bluesky.hyperion.experiment_plans import (
     pin_centre_then_xray_centre_plan,
     robot_load_then_centre_plan,
 )
-from mx_bluesky.hyperion.external_interaction.callbacks.common.callback_util import (
-    CallbacksFactory,
-    create_gridscan_callbacks,
-    create_load_centre_collect_callbacks,
-    create_robot_load_and_centre_callbacks,
-    create_rotation_callbacks,
-)
 from mx_bluesky.hyperion.parameters.gridscan import (
     GridScanWithEdgeDetect,
     HyperionSpecifiedThreeDGridScan,
@@ -47,44 +40,36 @@ class ExperimentRegistryEntry(TypedDict):
         | LoadCentreCollect
         | RobotLoadThenCentre
     ]
-    callbacks_factory: CallbacksFactory
 
 
 PLAN_REGISTRY: dict[str, ExperimentRegistryEntry] = {
     "flyscan_xray_centre": {
         "setup": flyscan_xray_centre_plan.create_devices,
         "param_type": HyperionSpecifiedThreeDGridScan,
-        "callbacks_factory": create_gridscan_callbacks,
     },
     "grid_detect_then_xray_centre": {
         "setup": grid_detect_then_xray_centre_plan.create_devices,
         "param_type": GridScanWithEdgeDetect,
-        "callbacks_factory": create_gridscan_callbacks,
     },
     "rotation_scan": {
         "setup": rotation_scan_plan.create_devices,
         "param_type": RotationScan,
-        "callbacks_factory": create_rotation_callbacks,
     },
     "pin_tip_centre_then_xray_centre": {
         "setup": pin_centre_then_xray_centre_plan.create_devices,
         "param_type": PinTipCentreThenXrayCentre,
-        "callbacks_factory": create_gridscan_callbacks,
     },
     "robot_load_then_centre": {
         "setup": robot_load_then_centre_plan.create_devices,
         "param_type": RobotLoadThenCentre,
-        "callbacks_factory": create_robot_load_and_centre_callbacks,
     },
     "multi_rotation_scan": {
         "setup": rotation_scan_plan.create_devices,
         "param_type": MultiRotationScan,
-        "callbacks_factory": create_rotation_callbacks,
     },
     "load_centre_collect_full": {
         "setup": load_centre_collect_full_plan.create_devices,
         "param_type": LoadCentreCollect,
-        "callbacks_factory": create_load_centre_collect_callbacks,
     },
 }
