+++ tests/unit_tests/hyperion/test_main_system.py
@@ -6,13 +6,11 @@ import os
 import threading
 from collections.abc import Callable
 from dataclasses import dataclass
-from queue import Queue
 from sys import argv
 from time import sleep
 from typing import Any
 from unittest.mock import MagicMock, patch
 
-import flask
 import pytest
 from blueapi.core import BlueskyContext
 from dodal.devices.attenuator.attenuator import BinaryFilterAttenuator
@@ -26,7 +24,6 @@ from mx_bluesky.hyperion.__main__ import (
     BlueskyRunner,
     Status,
     create_app,
-    create_targets,
     setup_context,
 )
 from mx_bluesky.hyperion.experiment_plans.experiment_registry import PLAN_REGISTRY
@@ -115,19 +112,10 @@ TEST_EXPTS = {
     "test_experiment": {
         "setup": MagicMock(),
         "param_type": MagicMock(),
-        "experiment_param_type": MagicMock(),
-        "callback_collection_type": MagicMock(),
-    },
-    "test_experiment_no_internal_param_type": {
-        "setup": MagicMock(),
-        "experiment_param_type": MagicMock(),
-        "callback_collection_type": MagicMock(),
     },
     "fgs_real_params": {
         "setup": MagicMock(),
         "param_type": HyperionSpecifiedThreeDGridScan,
-        "experiment_param_type": MagicMock(),
-        "callback_collection_type": MagicMock(),
     },
 }
 
@@ -235,10 +223,8 @@ def test_plan_with_no_params_fails(test_env: ClientAndRunEngine):
     ).json
     assert isinstance(response, dict)
     assert response.get("status") == Status.FAILED.value
-    assert (
-        response.get("message")
-        == "PlanNotFound(\"Corresponding internal param type for 'test_experiment_no_internal_param_type' not found in registry.\")"
-    )
+    assert isinstance(message := response.get("message"), str)
+    assert "'test_experiment_no_internal_param_type' not found in registry." in message
     test_env.mock_run_engine.abort()
 
 
@@ -388,81 +374,6 @@ def test_cli_args_parse(arg_list, parsed_arg_values):
     assert test_args.dev_mode == parsed_arg_values[0]
     assert test_args.verbose_event_logging == parsed_arg_values[1]
     assert test_args.skip_startup_connection == parsed_arg_values[2]
-    assert test_args.use_external_callbacks == parsed_arg_values[3]
-
-
-@patch("mx_bluesky.hyperion.__main__.do_default_logging_setup")
-@patch("mx_bluesky.hyperion.__main__.Publisher")
-@patch("mx_bluesky.hyperion.__main__.setup_context")
-@patch("dodal.log.GELFTCPHandler.emit")
-@patch("dodal.log.TimedRotatingFileHandler.emit")
-@pytest.mark.parametrize(["arg_list", "parsed_arg_values"], test_argument_combinations)
-def test_blueskyrunner_uses_cli_args_correctly_for_callbacks(
-    filehandler_emit,
-    graylog_emit,
-    setup_context: MagicMock,
-    zmq_publisher: MagicMock,
-    set_up_logging_handlers: MagicMock,
-    arg_list,
-    parsed_arg_values,
-):
-    mock_param_class = MagicMock()
-    callbacks_mock = MagicMock(
-        name="mock_callback_class",
-        return_value=("test_cb_1", "test_cb_2"),
-    )
-
-    TEST_REGISTRY = {
-        "test_experiment": {
-            "setup": MagicMock(),
-            "param_type": mock_param_class,
-            "callback_collection_type": callbacks_mock,
-        }
-    }
-
-    @dataclass
-    class MockCommand:
-        action: Actions
-        devices: Any = None
-        experiment: Any = None
-        parameters: Any = None
-        callbacks: Any = None
-
-    with (
-        flask.Flask(__name__).test_request_context() as flask_context,
-        patch("mx_bluesky.hyperion.__main__.Command", MockCommand),
-        patch.dict(
-            "mx_bluesky.hyperion.__main__.PLAN_REGISTRY", TEST_REGISTRY, clear=True
-        ),
-    ):
-        flask_context.request.data = b"{}"  # type: ignore
-        argv[1:] = arg_list
-        app, runner, port, dev_mode = create_targets()
-        runner.RE = MagicMock()
-        runner.command_queue = Queue()
-        runner_thread = threading.Thread(target=runner.wait_on_queue, daemon=True)
-        runner_thread.start()
-        assert dev_mode == parsed_arg_values[0]
-
-        mock_context = MagicMock()
-        mock_context.plan_functions = {"test_experiment": MagicMock()}
-        runner.command_queue.put(
-            MockCommand(
-                action=Actions.START,
-                devices={},
-                experiment="test_experiment",
-                parameters={},
-                callbacks=callbacks_mock,
-            ),  # type: ignore
-            block=True,  # type: ignore
-        )
-        runner.shutdown()
-        runner_thread.join()
-        assert (zmq_publisher.call_count == 1) == parsed_arg_values[3]
-        if parsed_arg_values[3]:
-            assert runner.RE.subscribe.call_count == 0
-        else:
-            assert runner.RE.subscribe.call_count == 2
 
 
 @pytest.mark.skip(
@@ -491,9 +402,7 @@ def test_when_blueskyrunner_initiated_then_plans_are_setup_and_devices_connected
         {
             "flyscan_xray_centre": {
                 "setup": fake_create_devices,
-                "run": MagicMock(),
                 "param_type": MagicMock(),
-                "callback_collection_type": MagicMock(),
             },
         },
         clear=True,
@@ -523,16 +432,14 @@ def test_when_blueskyrunner_initiated_and_skip_flag_is_set_then_setup_called_upo
         {
             "flyscan_xray_centre": {
                 "setup": mock_setup,
-                "run": MagicMock(),
                 "param_type": MagicMock(),
-                "callback_collection_type": MagicMock(),
             },
         },
         clear=True,
     ):
         runner = BlueskyRunner(MagicMock(), MagicMock(), skip_startup_connection=True)
         mock_setup.assert_not_called()
-        runner.start(lambda: None, test_fgs_params, "flyscan_xray_centre", None)
+        runner.start(lambda: None, test_fgs_params, "flyscan_xray_centre")
         mock_setup.assert_called_once()
         runner.shutdown()
 
@@ -545,26 +452,18 @@ def test_when_blueskyrunner_initiated_and_skip_flag_is_not_set_then_all_plans_se
             "flyscan_xray_centre": {
                 "setup": mock_setup,
                 "param_type": MagicMock(),
-                "experiment_param_type": MagicMock(),
-                "callback_collection_type": MagicMock(),
             },
             "rotation_scan": {
                 "setup": mock_setup,
                 "param_type": MagicMock(),
-                "experiment_param_type": MagicMock(),
-                "callback_collection_type": MagicMock(),
             },
             "other_plan": {
                 "setup": mock_setup,
                 "param_type": MagicMock(),
-                "experiment_param_type": MagicMock(),
-                "callback_collection_type": MagicMock(),
             },
             "yet_another_plan": {
                 "setup": mock_setup,
                 "param_type": MagicMock(),
-                "experiment_param_type": MagicMock(),
-                "callback_collection_type": MagicMock(),
             },
         },
         clear=True,
